requirements.txt
streamlit
pandas
fastapi
uvicorn[standard]
pydantic
python-multipart
requests
openai
python-dotenv
